requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
environ==1.0
Markdown==3.7
pillow==11.0.0
sqlparse==0.5.2
tzdata==2024.2
